moved classes from monolith window file to separate files
